debug: add logs on self_update

+++ src/includes/class-alma-plugin.php
@@ -61,42 +61,59 @@ class Alma_Plugin {
 	 * @return void
 	 */
 	protected function self_update() {
+		$this->logger->debug('self_update');
+
 		$db_version = get_option( 'alma_version' );
 
-		if ( version_compare( ALMA_VERSION, $db_version, '!=' ) ) {
+		if (
+			$db_version
+			&& version_compare( ALMA_VERSION, $db_version, '!=' )
+		) {
+			$this->logger->debug('old db version : ' . $db_version);
 
 			if (
 				$db_version
-				&& version_compare( $db_version, 4, '<' )
-				&& version_compare( ALMA_VERSION, 4, '>=' )
+				&& version_compare( $db_version, '4.0.0', '<' )
+				&& version_compare( ALMA_VERSION, '4.0.0', '>=' )
 			) {
 				$old_settings = get_option( 'woocommerce_alma_settings' );
 				update_option( Alma_Settings::OPTIONS_KEY, $old_settings );
+				$this->logger->debug('getting old settings');
 
 				// Upgrade to 4.
 				$gateway = new Alma_Payment_Gateway();
+				$this->logger->debug('initialize gateway');
 
 				// Manage credentials to match the new settings fields format.
 				try {
 					$gateway->manage_credentials();
+					$this->logger->debug('manage_credentials');
 				} catch ( \Exception $e ) { // phpcs:ignore Generic.CodeAnalysis.EmptyStatement.DetectedCatch
 					// We don't care if it fails there is nothing to update.
+					$this->logger->debug('the credentials were wrong');
 				}
 
 				if ( version_compare( $db_version, 3, '<' ) ) {
+					$this->logger->debug('Deactivate for version 2.*');
 					update_option( 'alma_version', ALMA_VERSION );
 					deactivate_plugins( 'alma-woocommerce-gateway/alma-woocommerce-gateway.php', true );
 				}
 			}
 
+			$this->logger->debug('Old plugin settings update and delete');
+
 			update_option( 'alma_version', ALMA_VERSION );
 			delete_option( 'woocommerce_alma_settings' );
 			delete_option( 'alma_warnings_handled' );
+			$this->logger->debug('Old plugin settings update and delete DONE');
 		}
 
 		if ( ! $db_version ) {
+			$this->logger->debug('No previous db version');
 			update_option( 'alma_version', ALMA_VERSION );
 		}
+
+		$this->logger->debug('self_update done');
 	}
 
 	/**
